Icon Enforcing & Widget Coloration Fix
Prefix text is now replaced with an image rather than a piece of text, as it was not possible to apply a custom font to a widget. The images used are converted '.png's of the Sawarabi Mincho font that my last commit added to the prefixes in the app's test window.

In addition, the images used to replace prefixes are now correctly tinted in the widget.